# Improve rendering/refresh time.