--- a/lib/vedeu/events/trigger.rb
+++ b/lib/vedeu/events/trigger.rb
@@ -83,6 +83,7 @@ module Vedeu
 
   # @!method trigger
   #   @see Vedeu::Events::Trigger.trigger
-  def_delegators Vedeu::Events::Trigger, :trigger
+  def_delegators Vedeu::Events::Trigger,
+                 :trigger
 
 end # Vedeu