--- a/lib/vedeu/geometry/dsl.rb
+++ b/lib/vedeu/geometry/dsl.rb
@@ -529,6 +529,7 @@ module Vedeu
 
   # @!method geometry
   #   @see Vedeu::Geometry::DSL.geometry
-  def_delegators Vedeu::Geometry::DSL, :geometry
+  def_delegators Vedeu::Geometry::DSL,
+                 :geometry
 
 end # Vedeu