--- a/lib/vedeu/groups/dsl.rb
+++ b/lib/vedeu/groups/dsl.rb
@@ -94,6 +94,7 @@ module Vedeu
 
   # @!method group
   #   @see Vedeu::Groups::DSL.group
-  def_delegators Vedeu::Groups::DSL, :group
+  def_delegators Vedeu::Groups::DSL,
+                 :group
 
 end # Vedeu