--- a/lib/vedeu/logging/all.rb
+++ b/lib/vedeu/logging/all.rb
@@ -12,5 +12,6 @@ end # Vedeu
 require 'vedeu/logging/lockless_log_device'
 require 'vedeu/logging/mono_logger'
 require 'vedeu/logging/log'
+require 'vedeu/logging/ips'
 require 'vedeu/logging/debug'
 require 'vedeu/logging/timer'