--- a/lib/vedeu/logging/ips.rb
+++ b/lib/vedeu/logging/ips.rb
@@ -1,0 +1,64 @@
+require 'benchmark/ips'
+
+module Benchmark
+  module IPS
+    class Report
+      class Entry
+        def display
+          Vedeu.log(type: :debug, message: to_s)
+        end
+      end
+    end
+  end
+end
+
+module Vedeu
+
+  module Logging
+
+    module Debug
+
+      class IPS
+
+        attr_accessor :samples
+        attr_accessor :benchmark
+
+        def initialize(&block)
+          @old_stdout = $stdout
+          $stdout     = StringIO.new
+          @samples    = {}
+          @benchmark  = Benchmark::IPS::Job.new
+          @count      = 0
+        end
+
+        def add_item(label = '', &blk)
+          samples[label] = blk
+          @count += 1
+
+          benchmark.item(label, &blk)
+        end
+
+        def execute!
+          benchmark.compare!
+          benchmark.run_warmup
+          benchmark.run
+
+          $stdout.sync = true
+          benchmark.run_comparison
+          benchmark.full_report
+
+          Vedeu.log(type: :debug, message: "IPS:\n#{$stdout.string}")
+          $stdout = @old_stdout
+
+          key = samples.keys.sample
+          Vedeu.log(type: :debug, message: "Running: #{key}")
+          samples[key].call
+        end
+
+      end # IPS
+
+    end # Debug
+
+  end # Logging
+
+end # Vedeu