--- a/lib/vedeu/output/output.rb
+++ b/lib/vedeu/output/output.rb
@@ -9,10 +9,29 @@ module Vedeu
     #
     class Output
 
+      # @param output [Array<Array<Vedeu::Views::Char>>|
+      #   NilClass|Vedeu::Models::Escape]
+      # @return [Array]
+      def self.buffer_write(output)
+        return nil if output.nil?
+
+        new(output).buffer_write
+      end
+
+      # @param output [Array<Array<Vedeu::Views::Char>>|
+      #   NilClass|Vedeu::Models::Escape]
+      # @return [Array<String>]
+      def self.direct_write(output)
+        return nil if output.nil?
+
+        new(output).direct_write
+      end
+
       # Writes output to the defined renderers.
       #
+      # @param output [Array<Array<Vedeu::Views::Char>>|
+      #   NilClass|Vedeu::Models::Escape]
       # @return [Array|NilClass|String]
-      # @see #initialize
       def self.render_output(output)
         return nil if output.nil?
 
@@ -21,10 +40,21 @@ module Vedeu
 
       # Return a new instance of Vedeu::Output::Output.
       #
-      # @param output [Array<Array<Vedeu::Views::Char>>]
+      # @param output [Array<Array<Vedeu::Views::Char>>|
+      #   NilClass|Vedeu::Models::Escape]
       # @return [Vedeu::Output::Output]
       def initialize(output)
         @output = output
+      end
+
+      # @return [Array]
+      def buffer_write
+        buffer_write!
+      end
+
+      # @return [Array<String>]
+      def direct_write
+        direct_write!
       end
 
       # Send the view to the renderers. If the output is a
@@ -81,6 +111,8 @@ module Vedeu
   # @!method render_output
   # @return [Array|NilClass]
   def_delegators Vedeu::Output::Output,
-                 :render_output
+                 :render_output,
+                 :buffer_write,
+                 :direct_write
 
 end # Vedeu