--- a/lib/vedeu/runtime/flags.rb
+++ b/lib/vedeu/runtime/flags.rb
@@ -63,7 +63,9 @@ module Vedeu
   # @!method ready?
   # @!method ready!
   # @return [Boolean]
-  def_delegators Vedeu::Runtime::Flags, :ready?, :ready!
+  def_delegators Vedeu::Runtime::Flags,
+                 :ready?,
+                 :ready!
 
   # :nocov: