--- a/lib/vedeu/runtime/router.rb
+++ b/lib/vedeu/runtime/router.rb
@@ -191,7 +191,8 @@ module Vedeu
 
   # @!method goto
   #   @see Vedeu::Runtime::Router#goto
-  def_delegators Vedeu::Runtime::Router, :goto
+  def_delegators Vedeu::Runtime::Router,
+                 :goto
 
   # :nocov: