… .rubocop.yml
… README.md
… docs/api.md
… docs/colours_styles.md
… docs/debugging.md
… docs/template.md
… test/lib/vedeu/output/output_test.rb
… test/lib/vedeu_test.rb
… test/test_helper.rb