Using an alternate mechanism to clear the content from a view, the rendering and refresh time has improved. From the 'examples/284_slow_rendering.rb' example:

**Old:**
Startup: ~350ms
Keypress: ~106ms

**New:**
Startup: ~245ms (~1.4 times faster)
Keypress: ~27ms (~3.9 times faster)